requirements.txt
aiohttp==3.8.4
aiohttp-middlewares==2.2.0
aioice==0.7.6
aiortc==1.2.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
av==8.1.0
black==23.3.0
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==38.0.4
dnspython==2.3.0
frozenlist==1.3.3
google-crc32c==1.5.0
h264-profile-level-id==1.0.0
idna==3.4
ifaddr==0.2.0
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pycparser==2.21
pydantic==1.10.10
pyee==8.2.2
Pygments==2.15.1
PyJWT==2.7.0
pylibsrtp==0.8.0
pymediasoup==0.2.2
pyOpenSSL==23.2.0
python-dotenv==1.0.0
sdp-transform==1.0.6
tomli==2.0.1
typing_extensions==4.7.0
websockets==11.0.3
yarl==1.9.2
